Add ERROR_CAMERA_REMOVED to CameraState (Internal Only)
This change introduces a new error code, ERROR_CAMERA_REMOVED,
to CameraState. This error signifies that a camera has been
unexpectedly removed from the system.

For now, this error code is restricted to internal library use
only and will not be exposed in the public Javadoc. This is a
temporary measure to allow for internal code changes as part
of a broader camera availability feature.

Once the complete camera availability functionality is ready
for review and public consumption, we will update the Javadoc
and officially expose ERROR_CAMERA_REMOVED for external use.

Bug: 427182232
Test: ./gradlew -p camera updateApi
Change-Id: I0df6cf0555bc5c9b364a216f729b90890cdfa2e3