Handle the null status of SpatialUser head for Headlocked UI test app
The head for the SpatialUser is in null status if it is not yet available. Once the user selects the "Head" radio button under the above situation, the test app doesn't handle this scenario and will crash. This patch deletes the not-null assertion operator applied to the head variable in SpatialUser.

Bug: 425737744
Test: On the Moohan device

Change-Id: I912eb1e1316fbc67ecb619375d0302a8cb9fae1d