Enforce androidx libraries do not enable coreLibraryDesugaringEnabled
coreLibraryDesugaringEnabled = true forces library users to also enable
core desugaring. This is a costly change and androidx libraries should
avoid doing that.

This change also removes coreLibraryDesugaringEnabled for
credentials/providerevents/providerevents-play-services library that
had it enabled by also moving to a newer version of gmscore that does
not require desugaring as that change was rolled back.

Test: ./gradlew projects -> passes
Bug: 428980475
Change-Id: I3c7a76b08a03483ad114992cb362cd8bd4d5c86c