Initialize Providers at Factory construction.
The current implementation of CameraDeviceSetupCompatFactory
waits until #getCameraDeviceSetupCompat() is called to try to create
Camera2CameraDeviceSetupCompatProvider and
PlayServiceCameraDeviceSetupCompatProvider. This meant that any Provider
specific intialization was delayed until the first call to
`#getCameraDevicesetupCompat`.

This CL moves the Provider creation to the
CameraDeviceSetupCompatFactory constructor to ensure that the Providers
are ready to go by the time #getCameraDeviceSetupCompat is called.

Bug: 349830046
Test: ./gradlew camera:featurecombinationquery:featurecombinationquery:assemble
Test: camera-testapp-featurecombo works as expected
Test: androidx.camera.integration.featurecombo.FeatureCombinationDeviceTest
      passes
Change-Id: I632635c9287337b38a2a4b329c524ef6d303b8b1