Context.receiveBroadcasts ktx extension.
This helper method is aimed to help Kotlin developers use coroutines to
handle their runtime registrations, which saves a few lines of code.

While this seems trivial to implement, there are a few race conditions
that are tricky to consider, which would easily hang a PendingResult
forever. That means that on top of saving a few lines of code, this
implementation helps with correctness, which is likely also an issue in
Java.

Change-Id: I7c668f4ebd1d2b57e30ac448ce86d1e70e3173c3