Skip switch to speaker if a BT endpoint is available
Users reported there calls switching from headphones
to speaker automatically. Upon investigation, when
the bluetooth device takes longer than a second to
connect, the core-telecom layer might switch the
video call to speaker if the first endpoint emitted
was earpiece.

To mitigate this unwanted behavior, if the core-telecom
layer detects a bluetooth endpoint, the assumption will
be made that the platform is in the process of connecting
to that endpoint and will skip auto-switching to speaker.

Bug: 427572739 (isolated switch to speaker ROOT bug)
Bug: 428529494 (secondary issue bug)
Test: 2 new unit tests + manual tests
Change-Id: I82463385ce37ef6d3f55f10f49702a3755cdb0bc