Enable dynamic updates in CameraCoordinator
This change enables both Camera2CameraCoordinator
and CameraCoordinatorAdapter to react to dynamic changes
in camera presence by implementing the
InternalCameraPresenceListener interface.

When the list of usable cameras changes, onCamerasUpdated()
is called. It re-evaluates the list of concurrent pairs
provided by the system and filters it to ensure that only
pairs where both cameras are present in the current usable
list AND are backward-compatible are exposed by CameraX.

Bug: 427182248
Test: Camera2CameraCoordinatorTest and CameraCoordinatorAdapterTest
Change-Id: I1e00fe6d56a8397ef08fff29a7e78157cb9d37ed