Add hashCode and equals to ImageResources
This will be needed for resources registering to ensure that ID
and version stays the same if resource hasn't changed.

For hashCode value, we will hash all fields (and compare in equal
method). This adds hash methods to all nested types that resources
are using. However, for simplicity and less initial development
effort, we will use byte array representation of the proto for
dynamic expressions. There is a follow up task to see how perfomant
this is and to optimize it further.

Bug: 428692423
Bug: 428693523
Test: Added
Relnote: "`ImageResource` now has `hashCode` and `equals` methods."
Change-Id: I650eece74a9494253f75839454bef2f35bb8c7e6