Refactor collection benchmark to separate Android target
Benchmark plugin does not suppport AGP KMP and we no longer want to support KMP projects applying com.android.library.

To support AGP KMP, we would need to bump the version of AGP in the benchmark plugin to 8.12.0-alpha8 (it is currently 8.1.1).  Also, in AGP KMP, we cannot set `androidTest.debugable = false`. This means that any consumers of the plugin would need to update their AGP version to match, which is a huge jump

Instead of forcing users to do that, I separated the existing project into iOS and android. In the android one, we can use the regular com.android.library without any issue as it is not KMP.

BUG: 428937955
TESTED: benchmarks run
Change-Id: I68810ae6f996395b2fbc15470a4fd060a301f63f