Consider that functions may be overloaded with differ upper bounds on type params
Ignoring type params and their upper bounds while collating functions means that overloads with different type params or upper bounds are considered to be the same function and are compared when they shouldn't be.

The downside of this is that actual incompatble changes to upper bounds are reported as removing old functions rather than changing them, but probably no good way around that.

Bug: 430572308
Test: Added a new test
Change-Id: I4b28b9a7bd85a36d47fda7375279a568e7bb4ce9