[scenecore] Allows opacity=0 with app passthrough enabled.
In the issue case, the app passthrough will be disabled by setting the PassthroughState to
PASSTHROUGH_MODE_OFF when opacity=0.

In this change, we allows the app opacity=0 with PassthroughState=PASSTHROUGH_MODE_MAX and
disabled the app passthrough when opacity=NO_PASSTHROUGH_OPACITY_PREFERENCE only.

Note that there is still an issue of "setting the app opacity to 0 will disable the passthrough"
in the underlying system.

Bug: 430111506
Test: scenecore/integration-tests/testapp
Change-Id: I746f8b91700d889e1808f210421f859cd89bfe14