Apply ImageCaptureFailedWhenVideoCaptureIsBoundQuirk to UniSoc chipset devices
On UniSoc chipset devices, apps can't take pictures successfully when the combination is VideoCapture + Preview + ImageCapture and all streams use 1280x720 resolution surfaces. Applying the quirk to force enable surface processing to workaround the issue.

Relnote: "Resolved an issue that caused devices with UniSoc chipsets to hang when capturing an image. The problem occurred when VideoCapture, Preview, and ImageCapture were used simultaneously with all streams set to a 1280x720 resolution."

Bug: 380802479
Test: manual test on real device
Change-Id: Ia00c4df593e44451ecc67726addbc4cf294c295d