Rely on callback for testing click.
Change the click test to rely on a callback to ensure that the click happened.
Current test which relies on starting Chrome activity shows a high
failure rate in post-submit tests.

Not starting Chrome for tests makes them faster and also removes the need
to navigate back after each test.

Bug: 424004315
Test: UserInteractionTests#ClickableRemoteView
Change-Id: I251aa53ef87a3d786fdcc6957f920cdbac13a82a