Adding a quick workaround for hit tests.
For some reason this helps forward facing rays, but it doesn't work in any direction. We need to revisit to we can have Whitebox work correctly.

Would like to submit this as it allows us to verify that the hit tests (and anchor generation) are actually working, while we sort out the ray direction in parallel.

Bug: 427961129
Change-Id: I8411c9831393337054282a9e15a13d362002c537