Fix DataBridgeClient#registerKeyUpdateCallback map
In DataBridgeClient#registerKeyUpdateCallback implementation, even when KeyUpdateCallbackWithExecutor is present in the list, a new redundant entry is added to the list and it leads to redundant callbacks when the registered keys are updated. Instead of adding it, check if an entry is already present, and only add if absent.

Bug: 432294119
Test: DataBridgeClientTest
Test: DataBridgeIntegrationTest
RelNote: NA
Change-Id: I844ec4989625bbc8e586825095e8d080935cae22