Remove disabling compose compiler plugin for AndroidXCompose projects
Prior to this change projects using AndroidXCompose plugin could disable
Compose compiler plugin. This was OK when it was originally disabled,
but over time projects such as ui-graphics added runtime dependency
without having the compiler in place missing on stability metadata as
seen in the bcv/native/current.txt changes.

Test: ./gradlew bOS --dry-run
Relnote: N/A
Bug: None
Change-Id: I8e12785b1365639b040cbe3f7d8e2d5c468e7844