Annotate AnchoredDraggable APIs with FrequentlyChangingValue
`offset`, `requireOffset` and `progress` change often and should not be read in composition.

Relnote: Annotated some AnchoredDraggable APIs with @FrequentlyChangingValue. `offset`, `requireOffset` and `progress` change often and should not be read in composition. Please access these values from the layout and draw phases, effects or otherwise outside of composition instead.
Test: ./gradlew :compose:foundation:foundation:foundation-samples:lint on a test sample that reads offset in composition
Change-Id: I0553910e71876e3354ee4f94080d5fecfd8cb032