Cache network capabilities to prevent unncessary IPCS
We cache the network capabilities to prevent unnnecessary IPCs to
connectivity manager. We should always have the up-to-date capabilities
since we get a callback whenever the network changes.

This can help with badly behaving apps that schedule/cancel very
frequently.

Bug: 427115602
Test: ./development/validate_changes.sh HEAD~1 buildOnServer,test --runOnDependentProjects
Relnote: "Cache network capabilities in SharedNetworkCallback to prevent
unnecessary IPCs"

Change-Id: Ie40279799b98b888bc15fbc51a9ca48d4c6b0bde