Fix: Color bug when using a vector drawable that uses one color and the fillColor is neither fully opaque nor fully transparent
- In case a vector drawable is using only one color, we avoid creating a bitmap with color information and only store the alpha for each pixel which acts as a mask. A color/tint filter is created using the same color as specified by the `fillColor` attrib in the XML and applied over the mask to produce the same output without the need for storing the color information for each pixel.
The issue was that the tint and the mask both have the alpha component,
effectively applying it twice. This has a visual effect only when an
alpha of anything but 0 or 1 is used in the `fillColor` attribute.

- Fixed by using an opaque tint filter as the opacity of the alpha mask still has an effect.

Relnote: Fixes a color bug when using a vector drawable that uses one color and the fillColor is neither fully opaque nor fully transparent
Fixes: 328677973
Test: Instrumentation test added
Change-Id: I3b04195622a5c24dab6e7f4f201dc7c27092a3d5