Fix ConcurrentModificationException in NavController
When we swapped NavControllerImpl over into the common source set and
out of the android source set, we lost access to the java
CopyOnWriteArrayList which was used by the list of destination changed
listeners. This made it possible to find a case where the list was being
modified while it was being iterated over.

We now interate over a copy of the list instead to present this error.

RelNote: "Fixed an error in NavController that caused a
ConcurrentModificationException when using
OnDestinationChangedListeners."
Test: modified test
Bug: 417784831

Change-Id: If7406ee832848e8f1fd06183f73801aff6dfcaed