Fix vertical padding in dialog buttons
According to the designer/spec, it should be 8dp.
However, buttons are 40dp visually, and expanded to
48dp from minimum interactive component size enforcement,
so there's extra logic added to ensure the visual padding
remains 8dp even if it's not actually 8dp in layout.

Also adds buttons to reopen dialog samples in catalog app.

Bug: n/a
Test: updated screenshot
Relnote: "Fixed button padding in AlertDialog when stacked vertically."
Change-Id: Ia211824e59881137b52bff81ad98ed848f6b899b