Reduce serialization cost of TimeFormatText.
Serializing a SimpleDateFormat instance ends up serializing a lot of
extra data that is not necessary to serialize for use on the same device
(e.g. names of weekdays, months, etc.). So instead of serializing the
SimpleDateFormat instance, serialize the pattern to the file, and
recreate the SimpleDateFormat from the serialized pattern when reading
back from the file.

Bug: 433949456
Change-Id: Ie79faf8edc56d3c2df670fde4a28b9c2c4130359