Use a different trace name when RecordingApplier is used
When Pausable Composition is enabled, there is a "Compose:applyChanges" block after the composition, but it doesn't actually perform apply changes. It uses a `RecordingApplier` to record the changes that will be applied when `PausedComposition.apply()` is called. It was confusing because the trace name was always "Compose:applyChanges" when Pausable Composition was enabled and when it was disabled.

In this change I'm renaming the trace to "Compose:recordChangesToApply" when `RecordingApplier` is being used. This will make it less confusing for people to analyze traces when they're not familiar with inner workings of `PausableComposition`.

Test: Manually captured a trace and confirmed that the trace name is changed only when `RecordingApplier` is used.
Relnote: n/a
Change-Id: I34d27520c0e03971dcc396deebd38f5e1b836df3