Made auto focus scroll tests more deterministic
I wasn't able to reproduce the failure, but after carefully reviewing
the parameters, I think the issue is that sometimes the focus line is
positioned right between items. This can cause different items to be
focused depending on the resolution. It's easy to avoid this for middle
positions, but it's much harder for corner cases when the focus line is
moving.

I've carefully updated all the test cases to make sure they don't put
the line between items. I've also added more logs in assert and changed
the way test case parameters are defined to make it easier to
investigate possible failures in the future. Also I set a constant
density to avoid rounding issues.

Fix: b/434055512
Test: GlimmerListAutoFocusScrollTest
Change-Id: I98ad69d03d288e403be52cff6983a2600b92af38