[Webkit] Remove names count check from MultiProfileTest
Up till no there's no way to unload profiles from memory, hence checking
on the getAllProfileNames won't be consistent as it may depend on any
loaded profiles in the process before running this test.

This was working just fine before as we didn't have any tests that load
the profile before, but that changed with
 OriginMatchedHeadersTest#settingHeaderOnlyAppliesToProvidedProfile
which loads a profile so removing the check so the tests could pass with
newer WebView versions.

Test: webkit/run_instrumentation_tests.sh
Change-Id: Iae8f9ba78ee87c182f199eba7442038954bc5102