Add sync triggers to the observed table states critical section.
In Iae612a1c4f8b8f1e8e15d2f60eb9bf68241570ef the invalidation tracker was updated to not start a transactions if there where no triggers to sync. This can lead to a race condition where one of two observers would not wait for trigger sync when `needsSync` is `true` causing one of the reactive return types to query the data too early and then never receive the invalidation. The transaction was acting as a lock mechanism, but it is too expensive so we want to avoid it.

This change iterates over the implementation of retrieving the tables to sync and resetting the flag by instead using a function that takes a lambda (onSync) where the lambda will be invoked around the ObservedTableStates's lock, recuperating the same behaviour as before but without using a transaction. Furthermore it enables us to keep the optimization of not asking for a connection nor starting a transaction if no sync is required.

Bug: 432365736
Test: FlowQueryTest
Change-Id: Ic9a3c7c3470578d08467841e552845b4787ca8af