Introduce AbstractNavigationEventInputHandler
Add AbstractNavigationEventInputHandler that users can extend to create their own input handlers.

It exposes a bunch of functions from the connected NavigationEventDispatcher. These functions are protected so that they are only visible in AbstractNavigationEventInputHandler or its subclasses.

The existing NavigationEventInputHandler now extends AbstractNavigationEventInputHandler. Its public functions can be used as a way to pass events from the outside.

This CL also makes NavigationEventDispatcher.addOnHasEnabledCallbacksChangedCallback internal again as it should only be useful in an input handler and is exposed there.

Bug: 432616296
Relnote: Introduce AbstractNavigationEventInputHandler
Test: existing tests still pass
Change-Id: I3897cdedbeabbb22174560595413e215b9237d95