Fix flaky Camera2DeviceSurfaceManagerTest
The Camera2DeviceSurfaceManagerTest is flaky because
of a race condition.

The test adds fake cameras using shadowCameraManager and
expects to receive specific camera updates. However, the
shadowCameraManager can also trigger its own camera
availability callbacks, which interferes with the test's
assertions.

This change makes the test more stable by removing some of
the verification logic that was sensitive to this race
condition. The updated test is less strict about the number
of callbacks received, which resolves the flakiness.

Bug: 434199701
Test: Camera2DeviceSurfaceManagerTest
Change-Id: I2dd3c244f718c8e1c844be2851662ec8e29e3e8d