Make the composer/composition internal interface abstract
This change makes many of the internal APIs used by the composition
to talk to the composer abstract. It prepares the way to add a different
implementation of the composer that uses a different slot table
representation to be added via a feature flag.

This changes cleans up some of the tests as adds a new version of
BitVector developed for the new slot table.

Relnote: N/A
Test: refactor, exisiting tests pass
Bug: 268366116
Change-Id: Ib8a9a4e2448525439a88f0a70d16a7efdbe4d466