Add AndroidXHostTestFailureHandlerPlugin to record test failures
This change introduces a new Gradle Settings plugin,
`AndroidXHostTestFailureHandlerPlugin`, that uses the Flow API
to record test failures to a signal file.

This is useful for CI systems or other tools to detect when a
test task has failed, even if the build continues due to the
`--continue` flag.

The plugin defines a `RecordTestFailureFlowAction` that writes
the failure information to `build/androidx_build_signals/test_task_failed_signal.txt`
in the settings directory.

The plugin is applied in the root `settings.gradle` file.

BUG: 370946159
Change-Id: I003ef4023d33ece8a9268f2e11ba25867cdb7a22