Graphics low latency: Fix buffer transforms for CanvasFrontBufferedRenderer and LowLatencyCanvasView
Bug: b/434190713, b/400986507

See aosp/2248293 for the initial implementation of this concept for GLFrontBufferedRenderer. When this
idea was ported to the Canvas-based implementations when they were first implemented, it seems like the
transforms were reversed. The documentation for the buffer transform hint says that the provided
transform is for the producer, while the inverse should be applied to the buffer that is consuming the
rendered content. In this case the producer is the renderer (SingleBufferedCanvasRenderer), while the
consumer is the SurfaceControlCompat.Transaction. Some variable names have been updated to reinforce
this fix.

GLFrontBufferedRenderer seems to be working fine in this regard.

Test: `./gradlew :graphics:graphics-core:install` then run `adb shell dumpsys SurfaceFlinger` while drawing - the "HWC layers" table should have 0 for the "Transform" column for the layers called "FrontBufferedLayer" and "MultiBufferedLayer", as opposed to `ROT_180`. And ideally the "Comp Type" column should be `DEVICE` rather than `CLIENT`
Change-Id: I36e55ff84b694fec843d595d9dd945c3f4b13991