Remove self reference optimization in compose agent
The ParameterFactory was implemented with a self reference
detection to avoid endless recursion. When a known reference was
found, we would not return the sub elements for the parameter but
just a reference to the previous expanded parameter.

This assumes the client could find the data for these reference.
Since we intend to use the same logic for state read values the
client (Android Studio) would not use the same cache and this
approach will not work.

Removing this feature allows us to remove the complexity of
storing cached value instances in ParameterFactory which probably
wasn't a good idea in the first place.

Bug: 435715387
Test: ParameterFactoryTest
Change-Id: Ibe6bcfc6ca43d8352d47dd53a6f081ed119c1556