Fix the issue that the initial anchor doesn't work
When creating PaneExpansionState for the first time, we were using a
default key to create its data, and then restoring it with the actual
data associated with the key. This worked fine until we make restoring
an async operaion, which makes the very initial measuring happen before
the restoring happens, therefore in the very first layout, the
calculated initial anchor offset is applied to the data of the default
key, instead of the actual key corresponding to the current scaffold
state.

The CL fixes this by using the provided key when creating the state
instance and its data.

It also:

1. Fixes a minor issue that when layout size changes, the dragging
   offset won't be correctly re-coerced
2. Adds more descriptive toString() to the PaneExpansionAnchor class.
3. Adds the usage of intitial anchor to our sample and tweaks demo
   anchor positions to make them more prominent.

Relnote: Fix the issue that initial anchors are not respected
Test: manual
Bug: 418296559
Change-Id: I32f5dac072a4870f669edab7cce6aafc7ce9467c