Fix errorprone for android projects
In r.android.com/3490363 we accidentally stopped setting up errorprone
because a call to finalizeDsl in onVariants does nothing (see
b/436595826). Instead, move to afterEvaluate that does run and allows us
to set up the task correctly.

In the original CL, I missed these tasks disappearing because I was
testing with `./gradlew runErrorProne --dry-run` which does not include
runErrorProneRelease, so I missed that these dissapeared.

Bug: 435264126
Test: ./gradlew webkit:webkit:runErrorProne
Change-Id: Iaec41bc503d680ded08267c3dc71ee84b25c839b