Fix LimitOffsetPagingSource not invalidating
The PagingSource currently registers database listeners upon first load. However, the progress of registering listeners, listener callback, and PagingSource invalidation is asynchronous. This means it is possible that database modifications in between the initial load and second load do not get propagatged in time to invalidate the PagingSource, leading to a possible stale second load.

Now the PagingSource registers listeners upon init to ensure it captures database modifications in time. However it doesn't invalidate the PagingSource unless initial refresh has completed. This is because:

1. Current PagingSource doesn't check for invalidation on Refresh load as well so this just maintains existing behaviour
2. Technically the states that are affected by database modifications are initialized during refresh load. So if refresh load hasn't been executed yet, database modifications would not lead to stale data.

Test setup is also updated to ensure that it waits for Paging's async tasks to complete before asserting Paging states.

Test: existing tests pass
Fixes: 422732937
Change-Id: Ida236d038049b7afca5f4d7865705eec4567600b