Add ability to customize VelocityTracker on platform level
Add the PlatformVelocityTracker interface to adjust the behavior of the
VelocityTracker on various platforms, such as filtering out unwanted
events. Current implementation of the VelocityTracker moved to the
DefaultVelocityTracker. Make VelocityPathFinder expect/actual so that it
can provide platform-specific behavior corresponding to the new
VelocityTracker.

Test: [DraggableTest], [VelocityPathFinderTest]

Relnote: "Adds the ability for VelocityTracker to calculate the velocity
of a pointer based on tracked pointer events using platform-specific
behavior."

Change-Id: I621e880bd3e77128aa546ec87e92fbff8c22f548