Use env var for remote cache push whose value does not change between postsubmit builds
The BuildCache interface does not support using Providers to set the `push` value.
In light of this, I've created an env var that abstracts the previous logic of setting `push` and set it in our busytown script.
The result of this is that configuration cache won't be invalidated due to BUILD_NUMBER changing when setting `push` (config cache is still invalidated but due to other issues)

BUG: 434944647
TESTED: ran IS_POSTSUBMIT=true ./gradlew tasks and saw that the uploaded build scan set remote push to true. set the var to false and the expected value was also set
Change-Id: Ie6c6bed64ee4426dacacec5320716179fe9da662