Re-enable isNonSuspendingPointerInputInDraggableEnabled
Re-enable the use of non suspending pointer input in draggable. This had been disabled due to a G3 breakage. It turns out there was a last step that needed to be taken into consideration, when a draggable is waiting for touchslop and something consumes during the final pass the draggable needs to give up the gesture.

Test: Added a new test.
Change-Id: I0c69f74f950a544a70d93550ade877999f386041